Refactor cabin management: implement custom hooks for create and delete operations, streamline CabinRow and CreateCabinForm components.
