+++ src/features/cabins/useCreateCabin.js
@@ -1,0 +1,20 @@
+import { useMutation, useQueryClient } from "@tanstack/react-query";
+
+import { createEditCabin } from "../../services/apiCabins";
+import { toast } from "react-hot-toast";
+
+export function useCreateCabin() {
+  const queryClient = useQueryClient();
+
+  const { isLoading: isCreating, mutate: createCabin } = useMutation({
+    mutationFn: createEditCabin,
+    onSuccess: () => {
+      toast.success("Cabin created successfully");
+      queryClient.invalidateQueries({ queryKey: ["cabins"] });
+    },
+    onError: (err) => {
+      toast.error(err.message);
+    },
+  });
+  return { isCreating, createCabin };
+}
